Move arena-ffi HTTP mapping unit tests out of mapping.rs into tests/playbook/http_mapping_spec.rs.
