+++ arena-ffi/BUILD
@@ -35,15 +35,6 @@ rust_shared_library(
 )
 
 rust_test(
-    name = "ffi_mapping_unit_test",
-    crate = ":arena_ffi",
-    size = "small",
-    edition = "2021",
-    deps = _ARENA_FFI_DEPS,
-    proc_macro_deps = _ARENA_FFI_PROC_MACRO_DEPS,
-)
-
-rust_test(
     name = "ffi_http_mapping_spec_test",
     srcs = ["tests/playbook/http_mapping_spec.rs"],
     size = "small",
